0.7.0 new sync methods

+++ src/mainwindow.cpp
@@ -360,7 +360,8 @@ void MainWindow::connectEvents()
     connect(Ramses::instance(),&Ramses::userChanged, this, &MainWindow::currentUserChanged);
     connect(Ramses::instance(), &Ramses::currentProjectChanged, this, &MainWindow::currentProjectChanged);
     connect(DBInterface::instance(),&DBInterface::connectionStatusChanged, this, &MainWindow::dbiConnectionStatusChanged);
-    connect(DBInterface::instance(), SIGNAL(synced()), this, SLOT(update()));
+    connect(DBInterface::instance(), SIGNAL(syncFinished()), this, SLOT(finishSync()));
+    connect(DBInterface::instance(), SIGNAL(syncStarted()), this, SLOT(startSync()));
 }
 
 void MainWindow::connectShortCuts()
@@ -818,6 +819,7 @@ void MainWindow::databaseSettingsAction()
 
 void MainWindow::home()
 {
+    mainToolBar->show();
     mainStack->setCurrentIndex(0);
 }
 
@@ -834,44 +836,56 @@ void MainWindow::revealUserFolder()
 
 void MainWindow::admin(bool show)
 {
+    mainToolBar->show();
     if (show) mainStack->setCurrentIndex(3);
     else home();
 }
 
 void MainWindow::projectSettings(bool show)
 {
+    mainToolBar->show();
     if (show) mainStack->setCurrentIndex(4);
     else home();
 }
 
 void MainWindow::pipeline(bool show)
 {
+    mainToolBar->show();
     if (show) mainStack->setCurrentIndex(5);
     else home();
 }
 
 void MainWindow::shots(bool show)
 {
+    mainToolBar->show();
     if (show) mainStack->setCurrentIndex(7);
     else home();
 }
 
 void MainWindow::assets(bool show)
 {
+    mainToolBar->show();
     if (show) mainStack->setCurrentIndex(6);
     else home();
 }
 
 void MainWindow::schedule(bool show)
 {
+    mainToolBar->show();
     if (show) mainStack->setCurrentIndex(8);
     else home();
 }
 
 void MainWindow::progress(bool show)
 {
-    if (show) mainStack->setCurrentIndex(9);
-    else home();
+    if (show) {
+        mainStack->setCurrentIndex(9);
+        mainToolBar->hide();
+    }
+    else {
+        home();
+        mainToolBar->show();
+    }
 }
 
 void MainWindow::install(bool show)
@@ -985,7 +999,6 @@ void MainWindow::freezeUI(bool f)
     if (f)
     {
         m_currentPageIndex = mainStack->currentIndex();
-        mainToolBar->hide();
         progress();
     }
     else
@@ -1032,11 +1045,23 @@ void MainWindow::dbiConnectionStatusChanged(NetworkUtils::NetworkStatus s)
     }
 }
 
-void MainWindow::fullSync()
+void MainWindow::finishSync()
 {
-    DBInterface::instance()->generalSync();
-    RamProject *proj = Ramses::instance()->currentProject();
-    if (proj) DBInterface::instance()->projectSync(proj->uuid());
+    if (m_closing) {
+        m_readyToClose = true;
+        this->close();
+        return;
+    }
+
+    ui_refreshButton->show();
+    mainStatusBar->showMessage(tr("Sync finished!"), 5000);
+}
+
+void MainWindow::startSync()
+{
+    ui_refreshButton->hide();
+    mainStatusBar->showMessage(tr("Syncing..."));
+
 }
 
 bool MainWindow::eventFilter(QObject *obj, QEvent *event)
@@ -1089,19 +1114,54 @@ bool MainWindow::eventFilter(QObject *obj, QEvent *event)
 
 void MainWindow::closeEvent(QCloseEvent *event)
 {
-    // Get to the home page first to make sure all toolbars are hidden
-    home();
+    if (m_closing && !m_readyToClose)
+    {
+        QMessageBox::StandardButton r = QMessageBox::question(this,
+                              tr("Closing Ramses..."),
+                              tr("I'm already closing, do you want me to force quit?\n\nThis may cause some data loss if the sync is not finished yet.")
+                              );
+        if (r == QMessageBox::Yes) m_readyToClose = true;
+    }
 
-    // Let's save the ui state
-    QSettings settings;
-    settings.beginGroup("ui");
-    settings.setValue("maximized", this->isMaximized());
-    settings.setValue("windowState", this->saveState());
-    settings.endGroup();
+    if (!m_readyToClose)
+    {
+        // Get to the home page first to make sure all toolbars are hidden
+        home();
 
-    QFontDatabase::removeAllApplicationFonts();
-    trayIcon->hide();
-    QMainWindow::closeEvent(event);
+        // Let's save the ui state
+        QSettings settings;
+        settings.beginGroup("ui");
+        settings.setValue("maximized", this->isMaximized());
+        settings.setValue("windowState", this->saveState());
+        settings.endGroup();
+
+        if (DBInterface::instance()->connectionStatus() == NetworkUtils::Online)
+        {
+            // Clean before quit!
+            m_closing = true;
+
+            ProgressManager *pm = ProgressManager::instance();
+            pm->setTitle("Disconnecting...");
+            pm->setText("One last sync!");
+            pm->setMaximum(3);
+            pm->start();
+            pm->freeze();
+
+            DBInterface::instance()->setOffline();
+
+            event->ignore();
+        }
+        else m_readyToClose = true;
+
+    }
+
+    if (m_readyToClose)
+    {
+        QFontDatabase::removeAllApplicationFonts();
+        trayIcon->hide();
+
+        QMainWindow::closeEvent(event);
+    }
 }
 
 void MainWindow::keyPressEvent(QKeyEvent *key)
